requirements.txt
Flask==2.1.2
tensorflow==2.11.0
numpy==1.21.0
Pillow==9.0.1
